# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Strata Palimpsest — AI memory vault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List all memories for the authenticated account
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query (title, content, tags)
        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: Filter by tag name
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [core, episode, concept, fact, emotion]
          description: Filter by memory type
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a new memory node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a single memory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /memories/{id}/pin:
    patch:
      operationId: toggleMemoryPin
      tags: [memories]
      summary: Toggle pinned status on a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Connections ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth schemas ─────────────────────────────────────────────────────────
...
    Memory:
      type: object
      required: [id, accountId, title, content, type, pinned, tags, createdAt, updatedAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [core, episode, concept, fact, emotion]
          description: core=foundational identity; episode=a past event; concept=an idea; fact=a known truth; emotion=a felt experience
        pinned:
          type: boolean
        tags:
          type: array
          items:
            type: string
        x:
          type: ["number", "null"]
          description: Graph canvas X position
        y:
          type: ["number", "null"]
          description: Graph canvas Y position
        sourceRef:
          type: ["string", "null"]
          description: External source URL or file reference
        instanceId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MemoryInput:
      type: object
      required: [title, content, type]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        type:
          type: string
          enum: [core, episode, concept, fact, emotion]
        tags:
          type: array
          items:
            type: string
        pinned:
          type: boolean
        x:
          type: number
        y:
          type: number
        sourceRef:
          type: ["string", "null"]
    MemoryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        type:
          type: string
          enum: [core, episode, concept, fact, emotion]
        tags:
          type: array
          items:
            type: string
        pinned:
          type: boolean
        x:
          type: number
        y:
          type: number
        sourceRef:
          type: ["string", "null"]
...